+++ src/apps/settings/modules/shared/store/app/StateBridge.ts
@@ -38,11 +38,12 @@ export const StaticSettingsToState = (
   userSettings: UserSettings,
   state: RootState,
 ): RootState => {
-  const { jsonSettings, yamlSettings, themes, layouts, placeholders } = userSettings;
+  const { jsonSettings, yamlSettings, themes, layouts, placeholders, wallpaper } = userSettings;
 
   return {
     ...state,
     ...jsonSettings,
+    wallpaper,
     availableThemes: themes,
     availableLayouts: layouts,
     availablePlaceholders: placeholders,