+++ src/apps/settings/modules/shared/store/app/reducer.ts
@@ -44,6 +44,7 @@ const initialState: RootState = {
     accent_lightest: '#000000',
     complement: null,
   },
+  wallpaper: null,
 };
 
 export const RootSlice = createSlice({