+++ src/background/exposed.rs
@@ -184,6 +184,8 @@ pub fn register_invoke_handler(app_builder: Builder<Wry>) -> Builder<Wry> {
         state_get_weg_items,
         state_get_settings,
         state_get_specific_apps_configurations,
+        state_get_wallpaper,
+        state_set_wallpaper,
         // Media
         media_prev,
         media_toggle_play_pause,