+++ src/background/modules/cli/domain.rs
@@ -5,6 +5,8 @@ use crate::state::domain::{Placeholder, Theme};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Resource {
     pub id: String,
+    /// Url of wallpaper
+    pub wallpaper: Option<String>,
     pub resources: ResourceItems,
 }